refactor: modularize UI and add tests
