+++ index.html
@@ -250,6 +250,6 @@
     <script src="compute.js" defer></script>
     <script src="csv.js" defer></script>
     <script src="pdf.js" defer></script>
-    <script src="app.js" defer></script>
+    <script type="module" src="ui.js"></script>
 </body>
 </html>
