+++ package.json
@@ -11,7 +11,9 @@
   "license": "ISC",
   "type": "commonjs",
   "devDependencies": {
-    "jest": "^30.1.1"
+    "@babel/preset-env": "^7.28.3",
+    "jest": "^30.1.1",
+    "jest-environment-jsdom": "^30.1.2"
   },
   "dependencies": {
     "jspdf": "^2.5.1"
